+++ i18n.js
@@ -1,9 +1,10 @@
 const NextI18Next = require('next-i18next').default
-const { localeSubpaths } = require('next/config').default().publicRuntimeConfig
+const {localeSubpaths} = require('next/config').default().publicRuntimeConfig
 const path = require('path')
 
 module.exports = new NextI18Next({
-  otherLanguages: ['nn', 'nb'],
-  localeSubpaths,
-  localePath: path.resolve('./public/static/locales'),
+    defaultLanguage: 'nb',
+    otherLanguages: ['nn', 'en'],
+    localeSubpaths,
+    localePath: path.resolve('./public/static/locales'),
 })
