+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@date-io/date-fns": "^2.10.6",
     "@material-ui/core": "^4.11.0",
+    "@material-ui/lab": "4.0.0-alpha.57",
     "@material-ui/pickers": "^3.2.10",
     "@pwabuilder/pwainstall": "^1.6.6",
     "@statisticsnorway/ssb-component-library": "^2.0.29",
