+++ src/components/common/dialog/modal.module.scss
@@ -10,6 +10,7 @@
   padding: 1rem;
   color: #fff;
   z-index: 100;
+  touch-action: none;
 }
 
 .fbuModalDialog {
