+++ src/components/layout/Layout.tsx
@@ -15,6 +15,8 @@ import { useLoader } from '../../hocs/globalLoader';
 import Loader from '../common/Loader';
 import { useRouter } from 'next/router';
 
+import styles from './styles/layout.module.scss';
+
 export const THREE_MINUTES = 3 * 60 * 1000;
 
 type LayoutProps = {
@@ -34,12 +36,19 @@ const Layout = (props: LayoutProps) => {
     const {loading} = useLoader();
 
     const [footerContent, updateFooterContent] = useState(null);
+    const [headerContent, updateHeaderContent] = useState(null);
+    const [showHeader, setShowHeader] = useState(true);
+    const [showFooter, setShowFooter] = useState(false);
 
     const setFooterContent = (node: ReactNode) => {
         console.log('Footer: updating content');
         updateFooterContent(node);
     };
 
+    const setHeaderContent = (node: ReactNode) => {
+        console.log('header: updating content');
+        updateHeaderContent(node);
+    };
 
     const {reset} = useIdleTimer({
         timeout: THREE_MINUTES,
@@ -90,28 +99,32 @@ const Layout = (props: LayoutProps) => {
     // console.log('----------- loading: ', loading);
 
     return (
-        <html lang="no">
-        <Head>
-            <meta charSet='utf-8'/>
-            <meta httpEquiv='X-UA-Compatible' content='IE=edge'/>
-            <meta name='viewport'
-                  content='width=device-width,initial-scale=1,minimum-scale=1,maximum-scale=1'/>
-            <meta name='description' content='Description'/>
-            <meta name='keywords' content='Keywords'/>
-            <title>SSB Forbruksundersøkelse 2021</title>
-            <link rel="manifest" href="/manifest.json" crossOrigin="use-credentials"/>
-            <script async src="https://unpkg.com/pwacompat" crossOrigin="anonymous"></script>
+        <div className={styles.mainContainer}>
+            <Head>
+                <meta charSet='utf-8'/>
+                <meta httpEquiv='X-UA-Compatible' content='IE=edge'/>
+                <meta name='viewport'
+                      content='width=device-width,initial-scale=1,minimum-scale=1,maximum-scale=1'/>
+                <meta name='description' content='Description'/>
+                <meta name='keywords' content='Keywords'/>
+                <title>SSB Forbruksundersøkelse 2021</title>
+                <link rel="manifest" href="/manifest.json" crossOrigin="use-credentials"/>
+                <script async src="https://unpkg.com/pwacompat" crossOrigin="anonymous"></script>
 
-            <link rel="icon" href='/favicon-16x16.png' type='image/png' sizes='16x16'/>
-            <link rel="icon" href='/favicon-32x32.png' type='image/png' sizes='32x32'/>
-            <link rel="apple-touch-icon" href="/apple-touch-icon.png"></link>
-            <meta name="theme-color" content="#0b2e13"/>
-        </Head>
-        <div>
+                <link rel="icon" href='/favicon-16x16.png' type='image/png' sizes='16x16'/>
+                <link rel="icon" href='/favicon-32x32.png' type='image/png' sizes='32x32'/>
+                <link rel="apple-touch-icon" href="/apple-touch-icon.png"></link>
+                <meta name="theme-color" content="#0b2e13"/>
+            </Head>
             <LayoutContext.Provider
-                value={{ footerContent, setFooterContent }}
+                value={{
+                    showHeader, showFooter,
+                    setShowHeader, setShowFooter,
+                    headerContent, setHeaderContent,
+                    footerContent, setFooterContent,
+                }}
             >
-                <Header siteTitle="Forbruk 2021" version="0.1" isOnline={isOnline}/>
+                {showHeader && <Header siteTitle="Forbruk 2021" version="0.1" isOnline={isOnline}/>}
                 <Workspace>
                     <SWHelper isOnline={isOnline} firstVisitWeb={firstVisitWeb}/>
                     {loading && <Loader/>}
@@ -136,10 +149,9 @@ const Layout = (props: LayoutProps) => {
                 /> */ ''}
                     {!isPWA() && props.children}
                 </Workspace>
-                <Footer/>
+                {showFooter && <Footer/>}
             </LayoutContext.Provider>
         </div>
-        </html>
     );
 };
