+++ src/components/purchases/NewItem.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent, useEffect, useRef, useState } from 'react';
+import { ChangeEvent, useContext, useEffect, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import {
     TextField,
@@ -7,9 +7,12 @@ import {
     InputAdornment,
     FormHelperText
 } from '@material-ui/core';
+import { NewPurchasesContext } from './NewPurchase';
+import useClickOutsideDetector from '../../hocs/useClickOutsideDetector';
 
 import formStyles from '../form/form.module.scss';
 import styles from './styles/addPurchase.module.scss';
+import Modal from '../common/dialog/Modal';
 
 type NewItemState = {
     name: string;
@@ -20,17 +23,22 @@ type NewItemState = {
 
 const NewItem = () => {
     const {t} = useTranslation('purchases');
+    const formRef = useRef(null);
     const nameFieldRef = useRef(null);
+    const [showPopup, setShowPopup] = useState(true);
+
     const [values, setValues] = useState<NewItemState>({
         name: '',
         qty: '1',
-        kr: '0',
-        cents: '00',
+        kr: '',
+        cents: '',
     });
 
     useEffect(() => {
-        nameFieldRef.current.focus();
-    }, []);
+        if (showPopup) {
+            nameFieldRef.current.focus();
+        }
+    }, [showPopup]);
 
     const updateValue = (val: keyof NewItemState) => (e: ChangeEvent<HTMLInputElement>) => {
         setValues({
@@ -40,77 +48,74 @@ const NewItem = () => {
     };
 
     return (
-        <div className={`${formStyles.fbuForm} ${styles.newItemForm}`}>
-            <h3>{t('addPurchase.newItem.title')}</h3>
-            <TextField
-                inputRef={nameFieldRef}
-                value={values.name}
-                label={t('addPurchase.newItem.name.label')}
-                placeholder={t('addPurchase.newItem.name.placeholder')}
-                onChange={updateValue('name')}
-                className={`${formStyles.fbuFormField} ${styles.itemName}`}
-            />
+        <Modal
+            show={showPopup}
+            onClose={() => {
+                if (values.name && values.qty && values.kr) {
+                    setShowPopup(false);
+                }
+            }}
+            closeText="Lagre"
+            onCancel={() => { setShowPopup(false); }}
+            cancelText="Avbryt"
+        >
+            <div
+                ref={(el) => {
+                    formRef.current = el;
+                }}
+                className={`${formStyles.fbuForm} ${styles.newItemForm}`}
+            >
+                <h3>{t('addPurchase.newItem.title')}</h3>
+                <TextField
+                    inputRef={nameFieldRef}
+                    value={values.name}
+                    required
+                    label={t('addPurchase.newItem.name.label')}
+                    placeholder={t('addPurchase.newItem.name.placeholder')}
+                    onChange={updateValue('name')}
+                    className={`${formStyles.fbuFormField} ${styles.itemName}`}
+                />
 
-            <div className={`${formStyles.fbuFormGroup} ${styles.itemQtyGroup}`}>
-                <div className={formStyles.fbuFormField}>
-                    <Input
+                <div className={`${formStyles.fbuFormGroup} ${styles.itemQtyGroup}`}>
+                    <TextField
                         placeholder={t('addPurchase.newItem.qty.placeholder')}
                         id="newItem-qty"
+                        label={t('addPurchase.newItem.qty.label')}
                         value={values.qty}
+                        required
                         onChange={updateValue('qty')}
                         aria-described-by="qty-helper-text"
-                        inputProps={{
-                            'aria-label': 'quantity'
+                        InputProps={{
+                            startAdornment: (
+                                <InputAdornment position="start">stk</InputAdornment>
+                            )
                         }}
-                        className={`${styles.itemQty}`}
+                        className={`${formStyles.fbuFormField} ${styles.itemQty}`}
                     />
-                    <FormHelperText
-                        id="qty-helper-text"
-                        className={formStyles.fbuFormFieldHelp}
-                    >
-                        {t('addPurchase.newItem.qty.label')}
-                    </FormHelperText>
-                </div>
-                <div className={formStyles.fbuFormField}>
-                    <Input
+
+                    <TextField
                         placeholder={t('addPurchase.newItem.price.kr.placeholder')}
                         id="newItem-kr"
+                        label={t('addPurchase.newItem.price.kr.label')}
                         value={values.kr}
+                        required
                         onChange={updateValue('kr')}
                         aria-described-by="kr-helper-text"
-                        inputProps={{
-                            'aria-label': 'price-kr'
-                        }}
-                        className={`${styles.itemKr}`}
+                        className={`${formStyles.fbuFormField} ${styles.itemKr}`}
                     />
-                    <FormHelperText
-                        id="kr-helper-text"
-                        className={formStyles.fbuFormFieldHelp}
-                    >
-                        {t('addPurchase.newItem.price.kr.label')}
-                    </FormHelperText>
-                </div>
-                <div className={formStyles.fbuFormField}>
-                    <Input
+
+                    <TextField
                         placeholder={t('addPurchase.newItem.price.cents.placeholder')}
                         id="newItem-cents"
+                        label={t('addPurchase.newItem.price.cents.label')}
                         value={values.cents}
                         onChange={updateValue('cents')}
                         aria-described-by="cents-helper-text"
-                        inputProps={{
-                            'aria-label': 'price-cents'
-                        }}
-                        className={`${styles.itemCents}`}
+                        className={`${formStyles.fbuFormField} ${styles.itemCents}`}
                     />
-                    <FormHelperText
-                        className={formStyles.fbuFormFieldHelp}
-                        id="cents-helper-text"
-                    >
-                        {t('addPurchase.newItem.price.cents.label')}
-                    </FormHelperText>
                 </div>
             </div>
-        </div>
+        </Modal>
     );
 };
