+++ src/pages/dashboard/dashboard.module.scss
@@ -4,15 +4,16 @@
   display: flex;
   flex-direction: column;
   width: 100%;
-  align-items: center;
+  height: 100%;
+  align-items: stretch;
   /* padding-top: 1rem; */
   background-color: $ssb-dark-1;
 }
 
 .floatingAddNew {
-  position: absolute;
-  right: 20px;
-  bottom: 100px;
+  position: absolute !important;
+  right: 2.5rem;
+  bottom: 5rem;
 
   a, li {
     background-color: #00824D !important;
